Warn about `download_slang_binaries` and `build_slang_from_source` features under Windows
The features work per-se, are useless for many use cases. E.g. if a crate's build script depends on Slang-rs, then the build will fail since Slang.dll will be in a place unknown to the compiled build script. Fixing this would require either

 (a) copying Slang.dll into the a-priory unknown `out` directory containing the build script, or

 (b) the $PATH seen by the build script would have to include the `bin` directory of the created Slang install tree.

Cargo provides no means of accomplishing this while staying in the realm of proper and portable usage. Even with ridiculous hackery it's next to impossible, certainly too unfeasible to even attempt.